Add Listing.hs
This CGI-program is able to list all existing books in the database. Intense research at the PROJECT-SOM-LABS reveals that some parts of the communication between database and program are different under apache-runtime conditions, so the program is very careful in those situations. It *should* be to use the right charsets in most of cases of usage. If not, check your .odbc.ini.